+++ packages/Bio/src/utils/monomer-lib/monomer-lib-base.ts
@@ -6,15 +6,18 @@ import wu from 'wu';
 import {Observable, Subject} from 'rxjs';
 
 import {IMonomerLibBase, Monomer, RGroup} from '@datagrok-libraries/bio/src/types/index';
-import {HelmAtom, HelmType, IWebEditorMonomer, MonomerType, PolymerType} from '@datagrok-libraries/bio/src/helm/types';
+import {HelmAtom, HelmType, IMonomerColors, IWebEditorMonomer, MonomerType, PolymerType} from '@datagrok-libraries/bio/src/helm/types';
 import {getMonomerHandleArgs} from '@datagrok-libraries/bio/src/helm/helm-helper';
 import {helmTypeToPolymerType} from '@datagrok-libraries/bio/src/monomer-works/monomer-works';
 import {HelmTypes, PolymerTypes} from '@datagrok-libraries/bio/src/helm/consts';
-import {HELM_REQUIRED_FIELD as REQ, HELM_RGROUP_FIELDS as RGP} from '@datagrok-libraries/bio/src/utils/const';
+import {HELM_OPTIONAL_FIELDS as OPT, HELM_REQUIRED_FIELD as REQ, HELM_RGROUP_FIELDS as RGP} from '@datagrok-libraries/bio/src/utils/const';
 import {GAP_SYMBOL, GapOriginals, NOTATION} from '@datagrok-libraries/bio/src/utils/macromolecule/consts';
+import {Vector} from '@datagrok-libraries/utils/src/type-declarations';
+import {vectorAdd, vectorDotProduct, vectorLength} from '@datagrok-libraries/utils/src/vector-operations';
 
 import {AmbiguousWebEditorMonomer, GapWebEditorMonomer, MissingWebEditorMonomer} from './web-editor-monomer-dummy';
 import {LibraryWebEditorMonomer} from './web-editor-monomer-of-library';
+import {naturalMonomerColors} from './monomer-colors';
 
 import {_package} from '../../package';
 
@@ -23,6 +26,10 @@ const monomerRe = /[\w()]+/;
 const ambMonomerRe = RegExp(String.raw`\(${monomerRe}(,${monomerRe})+\)`);
 
 export type MonomerLibDataType = { [polymerType: string]: { [monomerSymbol: string]: Monomer } };
+
+const whiteColorV = new Vector([255.0, 255.0, 255.0]);
+const blackColorV = new Vector([0.0, 0.0, 0.0]);
+const maxTextColorVLen = vectorLength(whiteColorV) * 0.7;
 
 export class MonomerLibBase implements IMonomerLibBase {
   protected _isEmpty: boolean;
@@ -143,7 +150,7 @@ export class MonomerLibBase implements IMonomerLibBase {
         resWem = m.wem = LibraryWebEditorMonomer.fromMonomer(biotype, m, this);
     }
 
-    return resWem!;
+    return resWem;
   }
 
   getTooltip(biotype: HelmType, monomerSymbol: string): HTMLElement {
@@ -155,10 +162,18 @@ export class MonomerLibBase implements IMonomerLibBase {
       const symbol = monomer[REQ.SYMBOL];
       const _name = monomer[REQ.NAME];
       const wem = this.getWebEditorMonomer(biotype, monomerSymbol)!;
+
       const htmlColor = wem.backgroundcolor;
       res.append(ui.divH([
-        ui.div([symbol], {style: {fontWeight: 'bolder', textWrap: 'nowrap', marginRight: '6px', color: htmlColor}}),
-        ui.div([monomer.name])
+        ui.div([symbol], {
+          style: {
+            /* fontWeight: 'bolder', */ textWrap: 'nowrap', marginLeft: '4px', marginRight: '4px',
+            color: wem.textcolor, backgroundColor: wem.backgroundcolor, borderColor: wem.linecolor,
+            borderWidth: '1px', borderStyle: 'solid', borderRadius: '2px', padding: '3px',
+            minWidth: '24px', textAlign: 'center',
+          }
+        }),
+        ui.div([monomer.name], {style: {padding: '4px'}}),
       ], {style: {display: 'flex', flexDirection: 'row', justifyContent: 'left'}}));
 
       // Structure
@@ -190,6 +205,77 @@ export class MonomerLibBase implements IMonomerLibBase {
     return res;
   }
 
+  getMonomerTextColor(biotype: HelmType, symbol: string): string {
+    const colors: IMonomerColors = this.getMonomerColors(biotype, symbol);
+    const htmlToA = (html: string): number[] => {
+      const rgbM = /rgb\(\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*\)/.exec(html);
+      if (rgbM)
+        return [parseInt(rgbM[1]), parseInt(rgbM[2]), parseInt(rgbM[3])];
+
+      const n = DG.Color.fromHtml(colors.textcolor!);
+      return [DG.Color.r(n), DG.Color.g(n), DG.Color.b(n)];
+    };
+
+    const textColorA = htmlToA(colors.textcolor!);
+    const textColorAdjV = new Vector([...textColorA.map((v) => v + 1)]);
+    const cosTextToWhite = vectorDotProduct(textColorAdjV, whiteColorV) /
+      (vectorLength(textColorAdjV) * vectorLength(whiteColorV));
+
+    const backColorA = htmlToA(colors.backgroundcolor!);
+    const backColorAdjV = new Vector([...backColorA.map((v) => v + 0.01)]);
+    const cosBackToWhite = vectorDotProduct(backColorAdjV, whiteColorV) /
+      (vectorLength(backColorAdjV) * vectorLength(whiteColorV));
+
+    let resColorA;
+    if (cosBackToWhite < cosTextToWhite)
+      resColorA = backColorA;
+    else
+      resColorA = textColorA;
+
+    let resColorV = new Vector(resColorA);
+    const resColorLen = vectorLength(resColorV);
+    if (resColorLen > maxTextColorVLen) resColorV = vectorAdd(blackColorV, resColorV, maxTextColorVLen / resColorLen);
+    return `rgb(${resColorV[0]}, ${resColorV[1]}, ${resColorV[2]})`;
+  }
+
+  getMonomerColors(biotype: HelmType, symbol: string): IMonomerColors {
+    const currentMonomerSchema = 'default';
+    let monomerSchema: string = currentMonomerSchema;
+
+    const polymerType = helmTypeToPolymerType(biotype);
+    const monomer = this.getMonomer(polymerType, symbol)!;
+    let res: any;
+    if (monomer) {
+      if (monomer.meta && monomer.meta.colors) {
+        const monomerColors: { [colorSchemaName: string]: any } = monomer.meta.colors;
+        if (!(currentMonomerSchema in monomerColors)) monomerSchema = 'default';
+        res = monomerColors[monomerSchema];
+      }
+
+      if (!res) {
+        const biotypeColors: { [symbol: string]: string } | undefined = naturalMonomerColors[biotype];
+        const nColor: string = biotypeColors?.[monomer.symbol];
+        if (nColor) {
+          const nTextColor = DG.Color.toHtml(DG.Color.getContrastColor(DG.Color.fromHtml(nColor)));
+          res = {textColor: nTextColor, lineColor: '#202020', backgroundColor: nColor};
+        }
+      }
+
+      const naSymbol: string | undefined = monomer[OPT.NATURAL_ANALOG];
+      if (!res && naSymbol) {
+        return this.getMonomerColors(biotype, naSymbol);
+      }
+    }
+
+    if (!res)
+      res = {textColor: "#202020", lineColor: "#202020", backgroundColor: "#A0A0A0"};
+
+    return {
+      textcolor: res.text ?? res.textColor,
+      linecolor: res.line ?? res.lineColor,
+      backgroundcolor: res.background ?? res.backgroundColor
+    } as IMonomerColors;
+  }
 
   getRS(smiles: string): { [r: string]: string } {
     const newS = smiles.match(/(?<=\[)[^\][]*(?=])/gm);