+++ packages/Bio/src/utils/monomer-lib/web-editor-monomer-of-library.ts
@@ -2,12 +2,11 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
-import {HelmType, IMonomerColors, IWebEditorMonomer, MonomerType, PolymerType, WebEditorRGroups} from '@datagrok-libraries/bio/src/helm/types';
+import {HelmType, IWebEditorMonomer, MonomerType, PolymerType, WebEditorRGroups} from '@datagrok-libraries/bio/src/helm/types';
 import {IMonomerLibBase, Monomer} from '@datagrok-libraries/bio/src/types/index';
 import {HELM_OPTIONAL_FIELDS as OPT, HELM_REQUIRED_FIELD as REQ, HELM_RGROUP_FIELDS as RGP} from '@datagrok-libraries/bio/src/utils/const';
 
-import {BrokenWebEditorMonomer, MissingWebEditorMonomer} from './web-editor-monomer-dummy';
-import {naturalMonomerColors} from './monomer-colors';
+import {BrokenWebEditorMonomer} from './web-editor-monomer-dummy';
 
 export class LibraryWebEditorMonomer implements IWebEditorMonomer {
   public get rs(): number { return Object.keys(this.at).length; }
@@ -59,7 +58,7 @@ export class LibraryWebEditorMonomer implements IWebEditorMonomer {
       monomer[REQ.MONOMER_TYPE],
       at);
 
-    const colors = getMonomerColors(biotype, monomer, monomerLib);
+    const colors = monomerLib.getMonomerColors(biotype, monomer[REQ.SYMBOL]);
     if (colors) {
       res.textcolor = colors?.textcolor;
       res.linecolor = colors?.linecolor;
@@ -68,42 +67,4 @@ export class LibraryWebEditorMonomer implements IWebEditorMonomer {
 
     return res;
   }
-}
-
-function getMonomerColors(biotype: HelmType, monomer: Monomer, monomerLib?: IMonomerLibBase): IMonomerColors | null {
-  const currentMonomerSchema = 'default';
-  let monomerSchema: string = currentMonomerSchema;
-
-  let res: any;
-  if (monomer.meta && monomer.meta.colors) {
-    const monomerColors: { [colorSchemaName: string]: any } = monomer.meta.colors;
-    if (!(currentMonomerSchema in monomerColors)) monomerSchema = 'default';
-    res = monomerColors[monomerSchema];
-  }
-
-  if (!res) {
-    const biotypeColors: { [symbol: string]: string } | undefined = naturalMonomerColors[biotype];
-    const nColor: string = biotypeColors?.[monomer.symbol];
-    if (nColor) {
-      const nTextColor = DG.Color.toHtml(DG.Color.getContrastColor(DG.Color.fromHtml(nColor)));
-      res = {textColor: nTextColor, lineColor: '#202020', backgroundColor: nColor};
-    }
-  }
-
-  const naSymbol: string | undefined = monomer[OPT.NATURAL_ANALOG];
-  if (!res && naSymbol) {
-    const polymerType = monomer[REQ.POLYMER_TYPE];
-    const naMonomer = monomerLib?.getMonomer(polymerType, naSymbol);
-    if (naMonomer)
-      return getMonomerColors(biotype, naMonomer);
-  }
-
-  if (!res)
-    res = {textColor: "#202020", lineColor: "#202020", backgroundColor: "#A0A0A0"};
-
-  return !res ? null : {
-    textcolor: res.text ?? res.textColor,
-    linecolor: res.line ?? res.lineColor,
-    backgroundcolor: res.background ?? res.backgroundColor
-  } as IMonomerColors;
 }